+++ src/main/java/com/booleanuk/api/cinema/movie/MovieController.java
@@ -1,13 +1,15 @@
 package com.booleanuk.api.cinema.movie;
 
+import com.booleanuk.api.cinema.response.ErrorResponse;
+import com.booleanuk.api.cinema.response.MovieListResponse;
+import com.booleanuk.api.cinema.response.MovieResponse;
+import com.booleanuk.api.cinema.response.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Date;
-import java.util.List;
 @RestController
 @RequestMapping("movies")
 public class MovieController {
@@ -16,51 +18,75 @@ public class MovieController {
     private MovieRepository movieRepository;
 
     @GetMapping
-    public List<Movie> getAllMovies() {
-        return this.movieRepository.findAll();
+    public ResponseEntity<MovieListResponse> getAllMovies() {
+        MovieListResponse movieListResponse = new MovieListResponse();
+        movieListResponse.set(this.movieRepository.findAll());
+        return ResponseEntity.ok(movieListResponse);
     }
 
     @PostMapping
-    public ResponseEntity<Movie> createMovie(@RequestBody Movie movie) {
+    public ResponseEntity<Response<?>> createMovie(@RequestBody Movie movie) {
+        MovieResponse movieResponse = new MovieResponse();
+
         if (movie.getTitle() == null || movie.getRating() == null || movie.getDescription() == null || movie.getRuntimeMins() == 0) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Some fields are invalid");
+            ErrorResponse error = new ErrorResponse();
+            error.set("Some fields are invalid");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+
+        } else {
+            Date today = new Date();
+            movie.setCreatedAt(today);
+            movie.setUpdatedAt(today);
+
+            movieResponse.set(this.movieRepository.save(movie));
         }
-
-
-
-        Date today = new Date();
-        movie.setCreatedAt(today);
-        movie.setUpdatedAt(today);
-
-        return new ResponseEntity<>(this.movieRepository.save(movie), HttpStatus.CREATED);
+        return new ResponseEntity<>(movieResponse, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Movie> updateMovie(@PathVariable int id, @RequestBody Movie movie) {
+    public ResponseEntity<Response<?>> updateMovie(@PathVariable int id, @RequestBody Movie movie) {
+        MovieResponse movieResponse = new MovieResponse();
+
         if (movie.getTitle() == null || movie.getRating() == null || movie.getDescription() == null || movie.getRuntimeMins() == 0) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Some fields are invalid");
+            ErrorResponse error = new ErrorResponse();
+            error.set("Some fields are invalid");
+            return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
+
         } else {
-            Movie movieToUpdate = this.movieRepository.findById(id).orElseThrow(
-                    () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found to update"));
+            Movie movieToUpdate = this.movieRepository.findById(id).orElse(null);
 
-            movieToUpdate.setTitle(movie.getTitle());
-            movieToUpdate.setRating(movie.getRating());
-            movieToUpdate.setDescription(movie.getDescription());
-            movieToUpdate.setRuntimeMins(movie.getRuntimeMins());
+            if (movieToUpdate == null) {
+                ErrorResponse error = new ErrorResponse();
+                error.set("Not found to update");
+                return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
 
-            Date today = new Date();
-            movieToUpdate.setUpdatedAt(today);
+            } else {
+                movieToUpdate.setTitle(movie.getTitle());
+                movieToUpdate.setRating(movie.getRating());
+                movieToUpdate.setDescription(movie.getDescription());
+                movieToUpdate.setRuntimeMins(movie.getRuntimeMins());
 
-            return new ResponseEntity<>(this.movieRepository.save(movieToUpdate), HttpStatus.CREATED);
+                Date today = new Date();
+                movieToUpdate.setUpdatedAt(today);
+                movieResponse.set(this.movieRepository.save(movieToUpdate));
+            }
+            return new ResponseEntity<>(movieResponse, HttpStatus.CREATED);
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Movie> deleteMovie(@PathVariable int id) {
-        Movie movieToDelete = this.movieRepository.findById(id).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found to delete"));
-        this.movieRepository.delete(movieToDelete);
+    public ResponseEntity<Response<?>> deleteMovie(@PathVariable int id) {
+        MovieResponse movieResponse = new MovieResponse();
+        Movie movieToDelete = this.movieRepository.findById(id).orElse(null);
 
-        return ResponseEntity.ok(movieToDelete);
+        if (movieToDelete == null) {
+            ErrorResponse error = new ErrorResponse();
+            error.set("Not found to delete");
+            return new ResponseEntity<>(error, HttpStatus.NOT_FOUND);
+        }
+        this.movieRepository.delete(movieToDelete);
+        movieResponse.set(movieToDelete);
+
+        return ResponseEntity.ok(movieResponse);
     }
 }
